pyproject.toml
[project]
name = "livekit-prec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.2",
    "duckduckgo-search>=8.1.1",
    "google-generativeai>=0.8.5",
    "groq>=1.0.0",
    "langchain-community>=0.4.1",
    "livekit-agents>=1.4.4",
    "livekit-murf>=0.1.1",
    "livekit-plugins-deepgram>=1.4.4",
    "livekit-plugins-google>=1.4.4",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.4.4",
    "livekit-plugins-silero>=1.4.4",
    "mcp>=1.26.0",
    "mem0ai>=1.0.1",
    "murf>=2.3.0",
    "opencv-python>=4.13.0.92",
    "pydantic-ai-slim[mcp,openai]>=1.66.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
